+++ .github/workflows/ci.yaml
@@ -136,6 +136,8 @@ jobs:
       DOCKER_SERVER: docker.io
       DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
       DOCKER_PASSWORD: ${{ secrets.DOCKER_PASSWORD }}
+      DOCKER_CI_USERNAME: ${{ secrets.DOCKERHUB_CI_USERNAME }}
+      DOCKER_CI_TOKEN: ${{ secrets.DOCKERHUB_CI_TOKEN }}
       AZURE_SAS_TOKEN: ${{ secrets.AZURE_SAS_TOKEN }}
       # Note: ${{ github.sha }} not working, see https://github.com/actions/checkout/issues/299
       NEURO_EXTRAS_IMAGE: image:neuromation/neuro-extras:${{ github.event.pull_request.head.sha
