+++ Makefile
@@ -20,7 +20,11 @@ test_e2e:
 	pytest -n 0                  ${PYTEST_FLAGS} -m "serial"     tests/e2e
 
 .PHONY: test
-test: lint test_e2e
+test: lint test_e2e test_unit
+
+.PHONY: test_unit
+test_unit:
+	pytest ${PYTEST_FLAGS} tests/unit
 
 .PHONY: changelog-draft
 changelog-draft:
