+++ setup.cfg
@@ -59,3 +59,6 @@ ignore_missing_imports = true
 
 [mypy-_pytest.*]
 ignore_missing_imports = true
+
+[mypy-deepdiff.*]
+ignore_missing_imports = true
