+++ tests/assets/registry_auths/default-registry.json
@@ -1,0 +1,7 @@
+{
+    "auths": {
+        "registry.default.org.neu.ro": {
+            "auth": "dGVzdC11c2VyOmRldGF1bHQtYXV0aAo="
+        }
+    }
+}
