+++ tests/assets/registry_auths/merged.json
@@ -1,0 +1,10 @@
+{
+    "auths": {
+        "registry.default.org.neu.ro": {
+            "auth": "dGVzdC11c2VyOmRldGF1bHQtYXV0aAo="
+        },
+        "https://index.docker.io/v1/": {
+            "auth": "dGVzdC11c2VyOmRvY2tlcmh1Yi1hdXRoCg=="
+        }
+    }
+}
